+++ CookMaster/app/src/main/java/com/example/cookmaster/model/Ingredient.java
@@ -1,7 +1,7 @@
 package com.example.cookmaster.model;
 
 public class Ingredient {
-    public Ingredient(long id, String name, double carbohydrates, double fats, double proteins, int kcal, double cost, String units){
+    public Ingredient(long id, String name, double carbohydrates, double fats, double proteins, double kcal, double cost, String units){
         this.id = id;
         this.name = name;
         this.carbohydrates = carbohydrates;
@@ -17,7 +17,7 @@ public class Ingredient {
     public double carbohydrates;
     public double fats;
     public double proteins;
-    public int kcal;
+    public double kcal;
     public double cost;
     public String units;
 }
